feat(PrecioHistorico): Change Double to BigDecimal with max 2 decimals

+++ rata-gamer/src/main/java/ar/edu/huergo/rata_gamer/dto/publicaciones/PrecioHistoricoDTO.java
@@ -1,6 +1,7 @@
 package ar.edu.huergo.rata_gamer.dto.publicaciones;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record PrecioHistoricoDTO(Long id, LocalDate fechaInicio, LocalDate fechaFin, Double precio) {
+public record PrecioHistoricoDTO(Long id, LocalDate fechaInicio, LocalDate fechaFin, BigDecimal precio) {
 }
